requirements.txt
customtkinter
apscheduler
google-generativeai
python-dotenv
requests
tweepy
